Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    int a;
-    scanf("%d",&a);
-    printf("%d",a);
+    float a;
+    scanf("%f",&a);
+    printf("%f",a);
     return 0;
 }
